+++ Maze.txt
@@ -1,0 +1,11 @@
+[][][][][][][][][][][]
+[]s []          []  []
+[]  []  []  []  []  []
+[]  []  []  []      []
+[]  [][][]  [][][]  []
+[]          []  []  []
+[][][][][][][]  []  []
+[]  []          []  []
+[]  []  []  [][][]  []
+[]      []         f[]
+[][][][][][][][][][][]
